% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_rad_mean}
\alias{WH_rad_mean}
\title{Circular mean}
\usage{
WH_rad_mean(theta)
}
\arguments{
\item{theta}{Numeric vector of radians.}
}
\value{
The circular mean in radians.
}
\description{
Computes the circular mean of a vector of radians.
}
\examples{
library(WH_circular)
y <- WH_HHMM_to_rad(c("23:00","23:15","23:30","23:45","00:00",
                    "01:00","01:15","01:30","01:45","02:00"))
print(WH_rad_to_HHMM(WH_mean_rad(y)))
}
